Enable automatic backend injection into tests
We currently use dynamically generated classes as the mechanism to have
per-backend tests using common code. While this works, it can be
slightly confusing as it's not very common practice. Now that we fully
utilise pytest for our tests, we could turn this dependency into a more
common pytest concept, and parametrise a "backend" parameter that would
be received by the test functions.

This test implements this parametrisation. Instead of requiring each
test to indicate this parametrisation, we implement it centrally.
Whenever a test function receives has a fixture called "backend" we
parametrise it with all the available backends.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>